feat: remove launch_ndt_convergence to ndt_convergence.py
